refactor: remove unused files and update project structure
Delete app.py, user model, authentication state, and related pages (login, register, dashboard)
Clean up assets by removing unused styles and favicon
Update .gitignore to include new patterns
Add new dependencies to requirements.txt
