+++ .gitignore
@@ -1,9 +1,9 @@
 .web
 assets/external/
 .states
+
 # Entorno virtual
 venv/
-.env
 
 # Archivos de Python
 __pycache__/
