+++ pacta/pages/register.py
@@ -1,13 +1,19 @@
 import reflex as rx
-from pacta_py.state.auth_state import AuthState
+from pacta.state.auth_state import AuthState
 
-def login():
+def register():
     return rx.center(
         rx.vstack(
-            rx.heading("PACTA - Login", size="1"),
+            rx.heading("PACTA - Registro", size="1"),
             rx.input(
                 placeholder="Username",
                 on_change=AuthState.set_username,
+                width="100%",
+                max_width="400px",
+            ),
+            rx.input(
+                placeholder="Email",
+                on_change=AuthState.set_email,
                 width="100%",
                 max_width="400px",
             ),
@@ -18,13 +24,9 @@ def login():
                 width="100%",
                 max_width="400px",
             ),
-            rx.checkbox(
-                "Recordarme",
-                on_change=AuthState.set_remember_me,
-            ),
             rx.button(
-                "Login",
-                on_click=AuthState.login,
+                "Registrar",
+                on_click=AuthState.register,
                 width="100%",
                 max_width="400px",
                 is_loading=AuthState.is_loading,
@@ -34,8 +36,8 @@ def login():
                 color="red",
             ),
             rx.link(
-                "¿No tienes cuenta? Regístrate aquí",
-                href="/register",
+                "¿Ya tienes cuenta? Inicia sesión aquí",
+                href="/login",
                 color="blue",
             ),
             spacing="4",
